move l10n mappings to i18n directory
