+++ src/app/hubs/hub-view/hub-properties-view/hub-properties-view.component.ts
@@ -5,7 +5,7 @@ import { MatCardModule } from '@angular/material/card';
 import { MatDividerModule } from '@angular/material/divider';
 import { NgIf } from '@angular/common';
 import { TranslocoModule } from '@ngneat/transloco';
-import { HUB_TYPE_TO_L10N_MAPPING } from '../../../mappings';
+import { HUB_TYPE_TO_L10N_MAPPING } from '../../../i18n';
 
 @Component({
     standalone: true,
