+++ src/app/i18n/index.ts
@@ -1,2 +1,6 @@
 export * from './provide-i18n';
 export * from './i18n-scopes';
+export * from './hub-type-to-l10n';
+export * from './io-type-to-l10n-key.pipe';
+export * from './io-operation-type-to-l10n-key.pipe';
+export * from './port-task-type-to-l10n';
