+++ src/app/i18n/io-type-to-l10n-key.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { IOType } from '../lego-hub';
-import { HUB_IO_I18N_SCOPE } from '../i18n';
+import { HUB_IO_I18N_SCOPE } from './index';
 
 @Pipe({
     name: 'ioTypeToL10nKey',
