--- src/app/mappings/index.ts
@@ -1,4 +1,0 @@
-export * from './hub-type-to-l10n';
-export * from './io-type-to-l10n-key.pipe';
-export * from './io-operation-type-to-l10n-key.pipe';
-export * from './port-task-type-to-l10n';
